Move transform functionality to separate file, fix partition type annotations
