+++ pytato/transform.py
@@ -25,7 +25,7 @@ THE SOFTWARE.
 """
 
 from typing import (Any, Callable, Dict, FrozenSet, Union, TypeVar, Set, Generic,
-                    List, Mapping, Iterable, Optional, Hashable, Tuple)
+                    List, Mapping, Iterable, Optional)
 
 from pytato.array import (
         Array, IndexLambda, Placeholder, MatrixProduct, Stack, Roll,
@@ -33,7 +33,7 @@ from pytato.array import (
         AbstractResultWithNamedArrays, Reshape, Concatenate, NamedArray,
         IndexRemappingBase, Einsum, InputArgumentBase,
         BasicIndex, AdvancedIndexInContiguousAxes, AdvancedIndexInNoncontiguousAxes,
-        IndexBase, make_placeholder)
+        IndexBase)
 from pytato.loopy import LoopyCall, LoopyCallResult
 from dataclasses import dataclass
 from pytato.tags import ImplStored
@@ -59,12 +59,9 @@ __doc__ = """
 .. autoclass:: CachedWalkMapper
 .. autoclass:: TopoSortMapper
 .. autoclass:: GraphToDictMapper
-.. autoclass:: CodePartitions
 .. autoclass:: CachedMapAndCopyMapper
 .. autofunction:: reverse_graph
 .. autofunction:: tag_child_nodes
-.. autofunction:: find_partitions
-.. autofunction:: execute_partitions
 .. autofunction:: copy_dict_of_named_arrays
 .. autofunction:: get_dependencies
 .. autofunction:: map_and_copy
@@ -1120,378 +1117,5 @@ def tag_child_nodes(graph: Dict[Array, Set[Array]], tag: Any,
 
 # }}}
 
-
-# {{{ graph partitioning
-
-class _GraphPartitioner(CachedMapper[ArrayOrNames]):
-    """Given a function *get_partition_id*, produces subgraphs representing
-    the computation. Users should not use this class directly, but use
-    :meth:`find_partitions` instead.
-    """
-
-    # {{{ infrastructure
-
-    def __init__(self, get_partition_id:
-                                   Callable[[Array], Hashable]) -> None:
-        super().__init__()
-
-        # Function to determine the Partition ID
-        self._get_partition_id = get_partition_id
-
-        # Naming for newly created PlaceHolders at partition edges
-        from pytools import UniqueNameGenerator
-        self.name_generator = UniqueNameGenerator(forced_prefix="_part_ph_")
-
-        # "edges" of the partitioned graph, maps an edge between two partitions,
-        # represented by a tuple of partition identifiers, to a set of placeholder
-        # names "conveying" information across the edge.
-        self.partition_pair_to_edges: Dict[Tuple[Hashable, Hashable],
-                Set[str]] = {}
-
-        self.var_name_to_result: Dict[str, Array] = {}
-
-        self._seen_node_to_placeholder: Dict[Array, Placeholder] = {}
-
-        # Reading the seen partition IDs out of partition_pair_to_edges is incorrect:
-        # e.g. if each partition is self-contained, no edges would appear. Instead,
-        # we remember each partition ID we see below, to guarantee that we don't
-        # miss any of them.
-        self.seen_partition_ids: Set[Hashable] = set()
-
-    def get_partition_id(self, array: Array):
-        part_id = self._get_partition_id(array)
-        self.seen_partition_ids.add(part_id)
-        return part_id
-
-    def does_edge_cross_partition_boundary(self, node1: Array, node2: Array) -> bool:
-        return self.get_partition_id(node1) != self.get_partition_id(node2)
-
-    def make_new_placeholder_name(self) -> str:
-        return self.name_generator()
-
-    def add_interpartition_edge(self, target: Array, dependency: Array,
-                                placeholder_name: str) -> None:
-        pid_target = self.get_partition_id(target)
-        pid_dependency = self.get_partition_id(dependency)
-
-        self.partition_pair_to_edges.setdefault(
-                (pid_target, pid_dependency), set()).add(placeholder_name)
-
-    def _handle_new_binding(self, expr: Array, child: Array) -> Any:
-        if self.does_edge_cross_partition_boundary(expr, child):
-            try:
-                ph = self._seen_node_to_placeholder[child]
-            except KeyError:
-                ph_name = self.make_new_placeholder_name()
-                # If an edge crosses a partition boundary, replace the
-                # depended-upon node (that nominally lives in the other partition)
-                # with a Placeholder that lives in the current partition. For each
-                # partition, collect the placeholder names that it’s supposed to
-                # compute.
-
-                ph = make_placeholder(ph_name,
-                    shape=child.shape,
-                    dtype=child.dtype,
-                    tags=child.tags)
-
-                self.var_name_to_result[ph_name] = self.rec(child)
-
-                self._seen_node_to_placeholder[child] = ph
-
-            assert ph.name
-            self.add_interpartition_edge(expr, child, ph.name)
-            return ph
-
-        else:
-            return self.rec(child)
-
-    def _handle_shape(self, expr: Array, shape: Any) -> Tuple[Any, ...]:
-        return tuple([
-            self._handle_new_binding(expr, dim) if isinstance(dim, Array) else dim
-            for dim in shape])
-
-    def __call__(self, expr: Array):
-        # Need to make sure the first node's partition is 'seen'
-        self.get_partition_id(expr)
-
-        return super().__call__(expr)
-
-    # }}}
-
-    # {{{ map_xxx methods
-
-    def map_named_array(self, expr: NamedArray) -> NamedArray:
-        return NamedArray(
-            container=self._handle_new_binding(expr, expr._container),
-            name=expr.name,
-            tags=expr.tags)
-
-    def map_index_lambda(self, expr: IndexLambda, *args: Any) -> IndexLambda:
-        return IndexLambda(expr=expr.expr,
-                shape=self._handle_shape(expr, expr.shape),
-                dtype=expr.dtype,
-                bindings={name: self._handle_new_binding(expr, child)
-                          for name, child in expr.bindings.items()},
-                tags=expr.tags)
-
-    def map_einsum(self, expr: Einsum, *args: Any) -> Einsum:
-        return Einsum(
-                     access_descriptors=expr.access_descriptors,
-                     args=tuple(self._handle_new_binding(expr, arg)
-                                for arg in expr.args),
-                     tags=expr.tags)
-
-    def map_matrix_product(self, expr: MatrixProduct, *args: Any) -> MatrixProduct:
-        return MatrixProduct(x1=self._handle_new_binding(expr, expr.x1),
-                             x2=self._handle_new_binding(expr, expr.x2),
-                             tags=expr.tags)
-
-    def map_stack(self, expr: Stack, *args: Any) -> Stack:
-        return Stack(
-                     arrays=tuple(self._handle_new_binding(expr, ary)
-                                  for ary in expr.arrays),
-                     axis=expr.axis,
-                     tags=expr.tags)
-
-    def map_concatenate(self, expr: Concatenate, *args: Any) -> Concatenate:
-        return Concatenate(
-                     arrays=tuple(self._handle_new_binding(expr, ary)
-                                  for ary in expr.arrays),
-                     axis=expr.axis,
-                     tags=expr.tags)
-
-    def map_roll(self, expr: Roll, *args: Any) -> Roll:
-        return Roll(array=self._handle_new_binding(expr, expr.array),
-                shift=expr.shift,
-                axis=expr.axis,
-                tags=expr.tags)
-
-    def map_axis_permutation(self, expr: AxisPermutation, *args: Any) \
-            -> AxisPermutation:
-        return AxisPermutation(
-                array=self._handle_new_binding(expr, expr.array),
-                axes=expr.axes,
-                tags=expr.tags)
-
-    def map_reshape(self, expr: Reshape, *args: Any) -> Reshape:
-        return Reshape(
-            array=self._handle_new_binding(expr, expr.array),
-            newshape=self._handle_shape(expr, expr.newshape),
-            order=expr.order,
-            tags=expr.tags)
-
-    def map_basic_index(self, expr: BasicIndex) -> BasicIndex:
-        return BasicIndex(
-                array=self._handle_new_binding(expr, expr.array),
-                indices=tuple(self._handle_new_binding(expr, idx)
-                                if isinstance(idx, Array) else idx
-                                for idx in expr.indices))
-
-    def map_contiguous_advanced_index(self,
-            expr: AdvancedIndexInContiguousAxes) -> AdvancedIndexInContiguousAxes:
-        return AdvancedIndexInContiguousAxes(
-                array=self._handle_new_binding(expr, expr.array),
-                indices=tuple(self._handle_new_binding(expr, idx)
-                                if isinstance(idx, Array) else idx
-                                for idx in expr.indices))
-
-    def map_non_contiguous_advanced_index(self,
-            expr: AdvancedIndexInNoncontiguousAxes) \
-            -> AdvancedIndexInNoncontiguousAxes:
-        return AdvancedIndexInNoncontiguousAxes(
-                array=self._handle_new_binding(expr, expr.array),
-                indices=tuple(self._handle_new_binding(expr, idx)
-                                if isinstance(idx, Array) else idx
-                                for idx in expr.indices))
-
-    def map_data_wrapper(self, expr: DataWrapper) -> DataWrapper:
-        return DataWrapper(
-                name=expr.name,
-                data=expr.data,
-                shape=self._handle_shape(expr, expr.shape),
-                tags=expr.tags)
-
-    def map_placeholder(self, expr: Placeholder, *args: Any) -> Placeholder:
-        assert expr.name
-
-        return Placeholder(name=expr.name,
-                shape=self._handle_shape(expr, expr.shape),
-                dtype=expr.dtype,
-                tags=expr.tags)
-
-    def map_size_param(self, expr: SizeParam) -> SizeParam:
-        assert expr.name
-        return SizeParam(name=expr.name, tags=expr.tags)
-
-    # }}}
-
-
-from pytato.target import BoundProgram
-
-
-@dataclass
-class CodePartitions:
-    """Store information about generated partitions.
-
-    .. attribute:: toposorted_partitions
-
-       List of topologically sorted partitions, represented by their
-       identifiers.
-
-    .. attribute:: partition_id_to_input_names
-
-       Mapping of partition identifiers to names of placeholders
-       the partition requires as input.
-
-    .. attribute:: partition_id_to_output_names
-
-       Mapping of partition IDs to the names of placeholders
-       they provide as output.
-
-    .. attribute:: var_name_to_result
-
-       Mapping of placeholder names to their respective :class:`pytato.array.Array`
-       they represent.
-    """
-    toposorted_partitions: List[Hashable]
-    partition_id_to_input_names: Dict[Hashable, Set[str]]
-    partition_id_to_output_names: Dict[Hashable, Set[str]]
-    var_name_to_result: Dict[str, Array]
-
-
-def find_partitions(outputs: DictOfNamedArrays,
-        part_func: Callable[[Array], Hashable]) ->\
-        CodePartitions:
-    """Partitions the *expr* according to *part_func* and generates code for
-    each partition.
-
-    :param expr: The expression to partition.
-    :param part_func: A callable that returns an instance of
-        :class:`Hashable` for a node.
-    :returns: An instance of :class:`CodePartitions` that contains the partitions.
-    """
-
-    pf = _GraphPartitioner(part_func)
-    rewritten_outputs = {name: pf(expr) for name, expr in outputs._data.items()}
-
-    partition_id_to_output_names: Dict[Hashable, Set[str]] = {
-        pid: set() for pid in pf.seen_partition_ids}
-    partition_id_to_input_names: Dict[Hashable, Set[str]] = {
-        pid: set() for pid in pf.seen_partition_ids}
-
-    partitions = set()
-
-    var_name_to_result = pf.var_name_to_result.copy()
-
-    for out_name, rewritten_output in rewritten_outputs.items():
-        out_part_id = part_func(outputs._data[out_name])
-        partition_id_to_output_names.setdefault(out_part_id, set()).add(out_name)
-        var_name_to_result[out_name] = rewritten_output
-
-    # Mapping of nodes to their successors; used to compute the topological order
-    partition_nodes_to_targets: Dict[Hashable, List[Hashable]] = {
-            pid: [] for pid in pf.seen_partition_ids}
-
-    for (pid_target, pid_dependency), var_names in \
-            pf.partition_pair_to_edges.items():
-        partitions.add(pid_target)
-        partitions.add(pid_dependency)
-
-        partition_nodes_to_targets[pid_dependency].append(pid_target)
-
-        for var_name in var_names:
-            partition_id_to_output_names.setdefault(
-                pid_dependency, set()).add(var_name)
-            partition_id_to_input_names.setdefault(
-                pid_target, set()).add(var_name)
-
-    from pytools.graph import compute_topological_order
-    toposorted_partitions = compute_topological_order(partition_nodes_to_targets)
-
-    result = CodePartitions(toposorted_partitions, partition_id_to_input_names,
-                          partition_id_to_output_names, var_name_to_result)
-
-    if __debug__:
-        _check_partition_disjointness(result)
-
-    return result
-
-
-class _SeenNodesWalkMapper(CachedWalkMapper):
-    def __init__(self) -> None:
-        super().__init__()
-        self.seen_nodes: Set[ArrayOrNames] = set()
-
-    def rec(self, expr: ArrayOrNames) -> None:  # type: ignore
-        super().rec(expr)
-        self.seen_nodes.add(expr)
-
-
-def _check_partition_disjointness(parts: CodePartitions):
-    part_id_to_nodes: Dict[Hashable, Set[ArrayOrNames]] = {}
-
-    for part_id, out_names in parts.partition_id_to_output_names.items():
-
-        mapper = _SeenNodesWalkMapper()
-        for out_name in out_names:
-            mapper(parts.var_name_to_result[out_name])
-
-        # check disjointness
-        for my_node in mapper.seen_nodes:
-            for other_part_id, other_node_set in \
-                    part_id_to_nodes.items():
-                assert (
-                    isinstance(my_node, Placeholder)
-                    or my_node not in other_node_set), (
-                        "partitions not disjoint: "
-                        f"{my_node.__class__.__name__} (id={id(my_node)}) "
-                        f"in both '{part_id}' and '{other_part_id}'")
-
-        part_id_to_nodes[part_id] = mapper.seen_nodes
-
-
-def generate_code_for_partitions(parts: CodePartitions) \
-        -> Dict[Hashable, BoundProgram]:
-    """Return a mapping of partition identifiers to their
-       :class:`pytato.target.BoundProgram`."""
-    from pytato import generate_loopy
-    prg_per_partition = {}
-
-    for pid in parts.toposorted_partitions:
-        d = DictOfNamedArrays(
-                    {var_name: parts.var_name_to_result[var_name]
-                        for var_name in parts.partition_id_to_output_names[pid]
-                     })
-        prg_per_partition[pid] = generate_loopy(d)
-
-    return prg_per_partition
-
-
-def execute_partitions(parts: CodePartitions, prg_per_partition:
-                        Dict[Hashable, BoundProgram], queue: Any) -> Dict[str, Any]:
-    """Executes a set of partitions on a :class:`pyopencl.CommandQueue`.
-
-    :param parts: An instance of :class:`CodePartitions` representing the
-        partitioned code.
-    :param queue: An instance of :class:`pyopencl.CommandQueue` to execute the
-        code on.
-    :returns: A dictionary of variable names mapped to their values.
-    """
-    context: Dict[str, Any] = {}
-    for pid in parts.toposorted_partitions:
-        # find names that are needed
-        inputs = {"queue": queue}
-
-        inputs.update({
-            k: context[k] for k in parts.partition_id_to_input_names[pid]
-            if k in context})
-
-        res = prg_per_partition[pid](**inputs)
-
-        context.update(res[1])
-
-    return context
-
-# }}}
 
 # vim: foldmethod=marker
